type change

+++ July2024/src/__init__.mojo
@@ -1,7 +1,7 @@
 from .matrix import Matrix
 from .test import test_matmul, bench_matmul
 
-alias Type = DType.float32
+alias Type = DType.float16
 alias TestSize = 1024
 alias BenchIters = 512
 alias MatmulSignature = fn[Type: DType, M: Int, N: Int, K: Int, //](inout Matrix[Type, M, N], Matrix[Type, M, K], Matrix[Type, K, N]) -> None
